Added rot_x, rot_y, rot_z that all take in a vertex and an angle (in degrees because intuition is key), and use matrix multiplication to output the vertex rotated about the given axis for the required amount of degrees. Also got rid of the render_order() function since it honestly wasn't making any meaningful additions to functionality.
I decided that I'll be removing all of the Save, Load, and Color related
things from the app for very similar reasoning. It simply doesn't add
anything at this time. Perhaps in a more sophisticated version or in a
different small application that I make. It's just easier to do what I
can instead of trying to do everything and none of it turns out well.